+++ src/model/PreviewUnstakeRequest.ts
@@ -19,11 +19,11 @@
  */
 
 /**
- * - CALL_TYPE_STANDARD: Evaluated at standard margin call evaluation time  - CALL_TYPE_URGENT: Evaluated in realtime  - CALL_TYPE_DEBIT: Evaluated at debit call evaluation time
+ * PreviewUnstakeRequest represents a request to preview an unstaking operation.
  */
-export enum XMCallType {
-  XmCallTypeUnspecified = 'XM_CALL_TYPE_UNSPECIFIED',
-  CallTypeStandard = 'CALL_TYPE_STANDARD',
-  CallTypeUrgent = 'CALL_TYPE_URGENT',
-  CallTypeDebit = 'CALL_TYPE_DEBIT',
-}
+export type PreviewUnstakeRequest = {
+  /**
+   * Amount to preview unstaking
+   */
+  amount: string;
+};